requirements.txt
django
django-tastypie
django-allauth
libmagic
requests
elasticsearch
urllib3